Material 클래스 데이터 관리 개선 및 함수 추가
`Material` 클래스의 데이터 관리 방식을 개선하고, `UpdateCBufferView` 함수를 추가하여 `m_cbufferValues`와 `m_dirtyCBs`를 효율적으로 관리하도록 수정했습니다.

- `Material.cpp`: `ApplyShaderParams` 함수에서 `m_dirtyCBs` 대신 `m_viewDirtyCBs`를 사용하고, `m_cbufferValues` 대신 `m_cbufferView`를 참조하도록 변경.
- `Material.cpp`: 새로운 함수 `UpdateCBufferView` 추가.
- `Material.h`: `UpdateCBufferView` 함수 및 관련 멤버 변수(`m_cbufferView`, `m_viewDirtyCBs`) 추가.
- `ProxyCommand.cpp`: `ProxyCommand::ProxyCommand(MeshRenderer* pComponent)`에 `UpdateCBufferView` 호출 추가.
- `UIGageTest.cpp`: `GetDurabilityRatio` 반환 타입을 `int`에서 `float`로 변경.
- `Main.json`: 일부 노드 제거, `selection` 값 초기화, `view` 속성 수정.

+++ Dynamic_CPP/Assets/Script/UIGageTest.cpp
@@ -43,7 +43,7 @@ void UIGageTest::Update(float tick)
 	//내구도 퍼센트 계산
 	if (m_weaponSlot)
 	{
-		int curDur = m_weaponSlot->GetDurabilityRatio();
+		float curDur = m_weaponSlot->GetDurabilityRatio();
 		percent = curDur;
 	}
 	else
